backend/requirements.txt
# API
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Banco de Dados
psycopg2-binary==2.9.9
sqlalchemy==2.0.35

# Processamento de Dados
pandas==2.2.3
numpy==2.1.1
scipy==1.14.1
pyarrow==17.0.0

# Geocodificação
geopandas==1.0.1
rapidfuzz==3.10.0
unidecode==1.3.8

# Machine Learning
scikit-learn==1.5.2
joblib==1.4.2
statsmodels==0.14.4

# Utilitários
python-dateutil==2.9.0

# Autenticação
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Testes
pytest==8.3.5
httpx==0.28.1
